requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.11.9
openai==1.82.0
tiktoken==0.7.0
pypdf==4.1.0

qdrant-client==1.6.9

llama-index-core==0.10.40
llama-index-embeddings-openai==0.1.9
llama-index-llms-openai==0.1.9
llama-index-readers-file==0.1.3
